requirements.txt
streamlit==1.30.0
streamlit-option-menu==0.3.6
numpy==1.24.3
pickle-mixin==1.2.0
scikit-learn==1.2.2
pandas==1.5.3
tensorflow==2.15.0
Pillow==9.4.0